// src/app/shared/models/customer.ts
export interface Customer {

  id?: number;

  firstname: string;

  lastname: string;

  email: string;

  phone: string;

  company?: string;

  address?: string;

  customerType: string;

  leadStatus?: string;

  leadSource?: string;

  status: string;

  createdAt?: string;

  updatedAt?: string;

}